* Lots of files: Getting rid of warning due to type conversion, unused 	  args and other such stuff.
	* convert/u8/ConvertUsecodeU8.h : Marking processes, and fixing a few 
	  miskates.
	  
	* graphics/ShapeInfo.h, graphics/TypeFlags.cpp : Added some new unknown
	  typeflags.
	  
	* kernel/Application.cpp, misc/Console.* : Clicking an item will return its
	  quality. Added ability to scroll console with Pageup and Pagedown.

	* world/Item.cpp : Changed the NPC classid detecion in callUsecodeEvent 
	  since npcs aren't the only items to have somehing in npcnum. Fixed up
	  I_getTypeFlag which was seriously broken. The new code is horrible
	  to look at.
	  
	* world/Item.h : Added FLG_MONSTER_NPC. But, I'm not sure if that is 
	  correct.
	  
	* world/ItemSorter.cpp : Added a few consts. A few other minor changes.
	
	* usecode/u8intrinsics.h : All unimplemented processes should now be dummy 
	  processes.

+++ filesys/IDataSource.h
@@ -31,7 +31,7 @@ class IDataSource
 		IDataSource() {}
 		virtual ~IDataSource() {}
 
-		virtual uint32 read1()=0;
+		virtual uint8 read1()=0;
 		virtual uint16 read2()=0;
 		virtual uint16 read2high()=0;
 		virtual uint32 read3()=0;
@@ -116,7 +116,7 @@ class IFileDataSource: public IDataSource
 	bool good() const { return in->good(); }
 
 	//	Read a byte value
-	virtual uint32 read1()
+	virtual uint8 read1()
 	{
 		return static_cast<uint8>(in->get());
 	}
@@ -214,7 +214,7 @@ public:
 
 	virtual ~IBufferDataSource() { }
 
-	virtual uint32 read1() {
+	virtual uint8 read1() {
 		uint8 b0;
 		b0 = *buf_ptr++;
 		return (b0);
